Introduce a load file wizard!
This is another big ugly commit. This introduces a new file loading
UI to the terminal-based Monitor. You can just drag and drop a file
onto the path TextField, and enter the address you want to load the
file into.

We also now introduce an "are you sure?" style dialog for when the
user wants to quit the application from the Monitor.

There is also now a global MonitorOutput class that contains a list
of the various texts being outputted by the commands as well as the
Monitor UI.